requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0 